Renames IoddScalerConverter to IoddScalerReader to synchronize the naming with IoddScalerWriter
