+++ src/Conversion/IoddConverter.cs
@@ -7,7 +7,7 @@ public class IoddConverter : IIoddDataConverter
     public object Convert(ParsableDatatype datatypeDef, ReadOnlySpan<byte> data) => datatypeDef switch
     {
         ParsableComplexDataTypeDef complexType => IoddComplexConverter.Convert(complexType, data),
-        ParsableSimpleDatatypeDef simpleType => IoddScalarConverter.Convert(simpleType, data),
+        ParsableSimpleDatatypeDef simpleType => IoddScalarReader.Convert(simpleType, data),
         _ => throw new NotImplementedException()
     };
 }
