+++ src/Tests/Conversion.Tests/IoddScalarConverterTests.cs
@@ -12,7 +12,7 @@ public class IoddScalarConverterTests
     public void CanConvert4BitPositiveInteger()
     {
         var typeDef = new ParsableSimpleDatatypeDef("intp", KindOfSimpleType.Integer, 4);
-        object result = IoddScalarConverter.Convert(typeDef, new byte[] { 0b0000_0100 });
+        object result = IoddScalarReader.Convert(typeDef, new byte[] { 0b0000_0100 });
         _ = result.Should().Be(4);
     }
 
@@ -21,7 +21,7 @@ public class IoddScalarConverterTests
     {
         var typeDef = new ParsableSimpleDatatypeDef("intp", KindOfSimpleType.Integer, 4);
 
-        object result = IoddScalarConverter.Convert(typeDef, new byte[] { 0b0000_1100 });
+        object result = IoddScalarReader.Convert(typeDef, new byte[] { 0b0000_1100 });
 
         _ = result.Should().Be(-4);
     }
@@ -30,7 +30,7 @@ public class IoddScalarConverterTests
     public void CanConvert17BitPositiveInteger()
     {
         var typeDef = new ParsableSimpleDatatypeDef("intp", KindOfSimpleType.Integer, 17);
-        object result = IoddScalarConverter.Convert(typeDef, new byte[] { 0b00000000, 0b01101110, 0b01011010 });
+        object result = IoddScalarReader.Convert(typeDef, new byte[] { 0b00000000, 0b01101110, 0b01011010 });
         _ = result.Should().Be(28250);
     }
 
@@ -38,7 +38,7 @@ public class IoddScalarConverterTests
     public void CanConvert17BitNegativeInteger()
     {
         var typeDef = new ParsableSimpleDatatypeDef("intp", KindOfSimpleType.Integer, 17);
-        object result = IoddScalarConverter.Convert(typeDef, new byte[] { 0b00000001, 0b10010001, 0b10100110 });
+        object result = IoddScalarReader.Convert(typeDef, new byte[] { 0b00000001, 0b10010001, 0b10100110 });
         _ = result.Should().Be(-28250);
     }
 
@@ -46,7 +46,7 @@ public class IoddScalarConverterTests
     public void CanConvert32BitNegativeInteger()
     {
         var typeDef = new ParsableSimpleDatatypeDef("intp", KindOfSimpleType.Integer, 32);
-        object result = IoddScalarConverter.Convert(typeDef, new byte[] { 0b11111110, 0b01010000, 0b11110000, 0b00001100 });
+        object result = IoddScalarReader.Convert(typeDef, new byte[] { 0b11111110, 0b01010000, 0b11110000, 0b00001100 });
         _ = result.Should().Be(-28250100);
         _ = result.Should().BeOfType<int>();
     }
@@ -55,7 +55,7 @@ public class IoddScalarConverterTests
     public void CanConvert33BitNegativeInteger()
     {
         var typeDef = new ParsableSimpleDatatypeDef("intp", KindOfSimpleType.Integer, 33);
-        object result = IoddScalarConverter.Convert(typeDef, new byte[] { 0b0000001, 0b11111110, 0b01010000, 0b11110000, 0b00001100 });
+        object result = IoddScalarReader.Convert(typeDef, new byte[] { 0b0000001, 0b11111110, 0b01010000, 0b11110000, 0b00001100 });
         _ = result.Should().Be(-28250100);
         _ = result.Should().BeOfType<long>();
     }
@@ -64,7 +64,7 @@ public class IoddScalarConverterTests
     public static void CanConvert12BitUInteger()
     {
         var typeDef = new ParsableSimpleDatatypeDef("intp", KindOfSimpleType.UInteger, 12);
-        object result = IoddScalarConverter.Convert(typeDef, new byte[] { 0b00000000, 0b0000_0100 });
+        object result = IoddScalarReader.Convert(typeDef, new byte[] { 0b00000000, 0b0000_0100 });
         _ = result.Should().Be(4);
         _ = result.Should().BeOfType<ushort>();
     }
@@ -73,7 +73,7 @@ public class IoddScalarConverterTests
     public static void CanConvert4BitUInteger()
     {
         var typeDef = new ParsableSimpleDatatypeDef("intp", KindOfSimpleType.UInteger, 4);
-        object result = IoddScalarConverter.Convert(typeDef, new byte[] { 0b0000_0100 });
+        object result = IoddScalarReader.Convert(typeDef, new byte[] { 0b0000_0100 });
         _ = result.Should().Be(4);
         _ = result.Should().BeOfType<byte>();
     }
@@ -82,7 +82,7 @@ public class IoddScalarConverterTests
     public static void CanConvert17BitUInteger()
     {
         var typeDef = new ParsableSimpleDatatypeDef("intp", KindOfSimpleType.UInteger, 17);
-        object result = IoddScalarConverter.Convert(typeDef, new byte[] { 0b00000001, 0b01101110, 0b01011010 });
+        object result = IoddScalarReader.Convert(typeDef, new byte[] { 0b00000001, 0b01101110, 0b01011010 });
         _ = result.Should().Be(93786);
         _ = result.Should().BeOfType<uint>();
     }
@@ -91,7 +91,7 @@ public class IoddScalarConverterTests
     public static void CanConvert48BitUInteger()
     {
         var typeDef = new ParsableSimpleDatatypeDef("intp", KindOfSimpleType.UInteger, 48);
-        object result = IoddScalarConverter.Convert(typeDef, new byte[] { 0b00000000, 0b00000000, 0b00000000, 0b00000001, 0b01101110, 0b01011010 });
+        object result = IoddScalarReader.Convert(typeDef, new byte[] { 0b00000000, 0b00000000, 0b00000000, 0b00000001, 0b01101110, 0b01011010 });
         _ = result.Should().Be(93786);
         _ = result.Should().BeOfType<ulong>();
     }
@@ -100,7 +100,7 @@ public class IoddScalarConverterTests
     public static void CanConvertPositiveFloat32()
     {
         var typeDef = new ParsableSimpleDatatypeDef("intp", KindOfSimpleType.Float, 32);
-        object result = IoddScalarConverter.Convert(typeDef, new byte[] { 0b00111111, 0b01000000, 0b00000000, 0b00000000 });
+        object result = IoddScalarReader.Convert(typeDef, new byte[] { 0b00111111, 0b01000000, 0b00000000, 0b00000000 });
         _ = result.Should().Be(0.75);
         _ = result.Should().BeOfType<float>();
     }
@@ -109,7 +109,7 @@ public class IoddScalarConverterTests
     public static void CanConvertNegativeFloat32()
     {
         var typeDef = new ParsableSimpleDatatypeDef("intp", KindOfSimpleType.Float, 32);
-        object result = IoddScalarConverter.Convert(typeDef, new byte[] { 0b10111111, 0b01000000, 0b00000000, 0b00000000 });
+        object result = IoddScalarReader.Convert(typeDef, new byte[] { 0b10111111, 0b01000000, 0b00000000, 0b00000000 });
         _ = result.Should().Be(-0.75);
         _ = result.Should().BeOfType<float>();
     }
@@ -119,7 +119,7 @@ public class IoddScalarConverterTests
     public static void CanConvertAsciiString()
     {
         var typeDef = new ParsableStringDef("intp", 32, StringTEncoding.ASCII);
-        object result = IoddScalarConverter.Convert(typeDef, "Hello"u8);
+        object result = IoddScalarReader.Convert(typeDef, "Hello"u8);
         _ = result.Should().Be("Hello");
         _ = result.Should().BeOfType<string>();
     }
@@ -128,7 +128,7 @@ public class IoddScalarConverterTests
     public static void CanConvertUtf8String()
     {
         var typeDef = new ParsableStringDef("intp", 32, StringTEncoding.UTF8);
-        object result = IoddScalarConverter.Convert(typeDef, "Hello"u8);
+        object result = IoddScalarReader.Convert(typeDef, "Hello"u8);
         _ = result.Should().Be("Hello");
         _ = result.Should().BeOfType<string>();
     }
